+++ docs/changelog.rst
@@ -17,6 +17,8 @@ Changelog for djangoSHOP
 * All Cascade plugins use ``GlossaryField`` instead of a list of ``PartialFormField`` s. This is
   much more "Djangonic", but requires djangocms-cascade version 0.11 or later.
 * All urlpatterns are compatible with configurations adding a final / to the request URL.
+* The URL for accessing an Order object, now uses the order number instead of it's primary key.
+
 
 0.9.2
 =====